chore: add JSON schema validation to setOptions and setCatalogue functions
